refactor to extract camera and renderer from scene class
