+++ src/scripts/renderer/canvas.js
@@ -1,0 +1,5 @@
+export const canvas = document.getElementById('vr-port');
+
+export const initialWidth = canvas.clientWidth;
+export const initialHeight = canvas.clientHeight;
+export const initialAspect = initialWidth / initialHeight;
