--- src/scripts/routes/home.js
@@ -1,12 +1,0 @@
-const mainTag = new HTMLMainElement();
-mainTag.innerHTML = `
-  <h1>WebXR Physics</h1>
-  <p>
-    This is an OSU Senior Capstone project for group 47. The goal is to 
-    demonstrate the abilities of the WebXR API by creating a sample project 
-    that can be referenced by others later. The project is sponsored by Intel.
-  </p>
-  <a href="/sim">Check it out in VR</a>
-`;
-
-export default mainTag;
