+++ src/styles/style.css
@@ -1,9 +1,10 @@
 @import url('https://fonts.googleapis.com/css?family=Merriweather|Roboto');
 
-header {
+#vr-port {
   height: 100vh;
-  position: relative;
+  width: 100vw;
 }
+
 header h1 {
   margin: 0;
   max-width: 100%;
@@ -13,48 +14,50 @@ header h1 {
   top: 50%;
   transform: translate(-50%, -50%);
 }
-header h1, header::after {
+
+header h1,
+header::after {
   color: white;
   text-shadow: black 0 0 10px;
 }
+
 @keyframes arrow-anim {
-	0% {
-		bottom: 20px;
-	}
-	100% {
-		bottom: 0;
-	}
+  0% {
+    bottom: 20px;
+  }
+  100% {
+    bottom: 0;
+  }
 }
+
 header::after {
   animation: arrow-anim 1s infinite alternate ease-out;
-	font-weight: bold;
-	left: 50%;
-	position: absolute;
-	display: block;
-	font-size: 36px;
-	content: "￬";
-	padding: 10px;
-	transform: translateX(-50%);
-	z-index: 2;
-}
-canvas {
-  top: 0;
-  left: 0;
+  font-weight: bold;
+  left: 50%;
   position: absolute;
-  /* The renderer must be overriding these settings for some reason. */
-  width: 100% !important;
-  height: 100% !important;
-  z-index: -1;
+  display: block;
+  font-size: 36px;
+  content: '￬';
+  padding: 10px;
+  transform: translateX(-50%);
+  z-index: 2;
 }
 
 body {
-    margin:0px;
-    font-family: "Roboto", sans-serif;
-    background-color: #eee;
+  margin: 0px;
+  font-family: 'Roboto', sans-serif;
+  background-color: #eee;
 }
-h1, h2, h3, h4, h5, h6 {
-  font-family: "Merriweather", serif;
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  font-family: 'Merriweather', serif;
 }
+
 main {
   background-color: #fff;
   margin: 0 auto;
